process_policy_test.cc: Fix mixed-sign comparison warning.
Clang builds were failing like this:

In file included from ..\..\sandbox\win\src\process_policy_test.cc:19:
..\..\testing/gtest/include/gtest/gtest.h(1392,16):
error: comparison of integers of different signs: 'const unsigned long' and 'const int' [-Werror,-Wsign-compare]
  if (expected == actual) {
      ~~~~~~~~ ^  ~~~~~~
..\..\testing/gtest/include/gtest/gtest.h(1422,12):
note: in instantiation of function template specialization 'testing::internal::CmpHelperEQ<unsigned long, int>' requested here
    return CmpHelperEQ(expected_expression, actual_expression, expected,
           ^
..\..\sandbox\win\src\process_policy_test.cc(419,3):
note: in instantiation of function template specialization 'testing::internal::EqHelper<false>::Compare<unsigned long, int>' requested here
  EXPECT_EQ(STATUS_BREAKPOINT, runner.RunTest(L"Process_Crash"));
  ^
..\..\testing/gtest/include/gtest/gtest.h(1928,67):  note: expanded from macro 'EXPECT_EQ'
                      EqHelper<GTEST_IS_NULL_LITERAL_(expected)>::Compare, \
                                                                  ^

BUG=82385
TBR=wfh

Review URL: https://codereview.chromium.org/1833863004 .

Cr-Commit-Position: refs/heads/master@{#383286}

Former-commit-id: 035459168ac247f089c25d5f5c3c1f36814bb1f5

+++ sandbox/win/src/process_policy_test.cc
@@ -416,7 +416,8 @@ TEST(ProcessPolicyTest, CreateProcessAW) {
 // Tests that the broker correctly handles a process crashing within the job.
 TEST(ProcessPolicyTest, CreateProcessCrashy) {
   TestRunner runner;
-  EXPECT_EQ(STATUS_BREAKPOINT, runner.RunTest(L"Process_Crash"));
+  EXPECT_EQ(static_cast<int>(STATUS_BREAKPOINT),
+            runner.RunTest(L"Process_Crash"));
 }
 
 TEST(ProcessPolicyTest, OpenToken) {
